# res/openapi.yaml
openapi: 3.0.0
info:
  title: grommunio Admin API
  version: 1.0.0
servers:
  - url: '/api/v1'
    description: Default deployment path
paths:
...
  /passwd:
    put:
      summary: Change current users password
      tags:
        - Misc
        - Service
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [new, old]
              properties:
                new:
                  type: string
                  description: The new password
                old:
                  type: string
                  description: The old password
                user:
                  type: string
                  description: Set password for a specific user (use inline authentication)
      responses:
        '200':
          description: Password updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          description: Authentication failed
        '403':
          description: Incorrect old password
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /service/syncPolicy/{username}:
    get:
      summary: Get sync policy for specific user
      tags:
        - Service
      security:
        - JWTCookie: []
      parameters:
        - name: username
          in: path
          required: true
          description: E-mail address of the user
          schema:
            type: string
      responses:
        '200':
          description: Policy returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/syncPolicy'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
...
  /system/domains/{domainID}:
    get:
      summary: Get detailed info about domain
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/verbosity'
      responses:
        '200':
          description: Domain returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    patch:
      summary: Update domain
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/domain'
      responses:
        '200':
          description: Domain updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/domain'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    delete:
      summary: Delete domain
      description: Removes domain (purge=true) or marks it as deleted (purge=false)
      tags:
        - System Admin/Domains
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
        - name: purge
          in: query
          required: false
          description: Permanently delete the domain
          schema:
            type: boolean
            default: false
        - name: deleteFiles
          in: query
          required: false
          description: Delete files from disk when purging
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Domain deleted
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
...
  /domains/{domainID}/users:
    get:
      summary: Get lists of users
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/verbosity'
        - $ref: '#/components/parameters/queryLimit'
        - $ref: '#/components/parameters/queryOffset'
        - $ref: '#/components/parameters/match'
        - $ref: '#/components/parameters/matchFields'
        - name: sort
          description: Sort by attribute. Can be given multiple times.
          in: query
          schema:
            type: string
            pattern: '^(ID|username|domainID|pop3_imap|smtp|changePassword|publicAddress)(,(a|de)sc)?$'
        - $ref: '#/components/parameters/filterID'
        - name: username
          description: Filter by username
          in: query
          schema:
            $ref: '#/components/schemas/intList'
        - $ref: '#/components/parameters/propnames'
      responses:
        '200':
          description: Data returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    post:
      summary: Create new user
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userInit'
      responses:
        '201':
          description: Domain created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /domains/{domainID}/users/{userID}:
    get:
      summary: Get information about a specific user
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/verbosity'
      responses:
        '200':
          description: User data returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    patch:
      summary: Update user
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userUpdate'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    delete:
      summary: Delete user
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
        - name: deleteFiles
          description: Delete user files on disk
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: User deleted
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /domains/{domainID}/users/{userID}/delegates:
    get:
      summary: Get list of users that can send mails on the users behalf
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
      responses:
        '200':
          description: List of users returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: string
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    put:
      summary: Set list of users that can send mails on the users behalf
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: List updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /domains/{domainID}/users/{userID}/storeProps:
    get:
      summary: Get user store properties
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/propnames'
      responses:
        '200':
          description: Property values returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/propvals'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    patch:
      summary: Update store properties
      tags:
       - Domain Admin/Users
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
      requestBody:
        content:
          application/json:
           schema:
            type: object
            description: Tag name / value mapping of proptags to modify
            additionalProperties:
              oneOf:
                - type: integer
                - type: string
      responses:
        '200':
          description: Modification request successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                    description: Tag name / error message for tags that failed to update
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
    delete:
      summary: Delete user store props
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
        - $ref: '#/components/parameters/propnames'
      responses:
        '200':
          description: Properties removed
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /domains/{domainID}/users/{userID}/sync:
    get:
      summary: Get status of synced devices
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
      responses:
        '200':
          description: Devices returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    description: List of devices with their sync entries
                    items:
                      type: object
                      properties:
                        devicetype:
                          type: string
                        useragent:
                          type: string
                        firstsynctime:
                          type: integer
                          description: UNIX timestamp of first sync
                        lastupdatetime:
                          type: integer
                          description: UNIX timestamp of last update
                        asversion:
                          type: string
                          description: Sync version
                        foldersSyncable:
                          type: integer
                          description: Number of folders that can be synced
                        foldersSynced:
                          type: integer
                          description: Number of folders that were actually synced
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
  /domains/{domainID}/users/{userID}/sync/{ID}:
    delete:
      summary: Remove device (issue a resync)
      tags:
        - Domain Admin/Users
      security:
        - JWTCookie: []
      parameters:
        - $ref: '#/components/parameters/userID'
        - $ref: '#/components/parameters/domainID'
        - name: ID
          in: path
          required: true
          description: Device ID
          schema:
            type: string
      responses:
        '200':
          description: Device resync queued
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/DatabaseError'
...
components:
  securitySchemes:
    JWTCookie:
      type: apiKey
      in: cookie
      name: grommunioAuthJwt
  parameters:
...
    domainID:
      name: domainID
      description: ID of the domain
      in: path
      required: true
      schema:
        type: integer
...
    verbosity:
      name: level
      in: query
      description: Set detail level of return value. Usually ranges from 0 to 2.
      schema:
        type: integer
    queryLimit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        default: 50
    queryOffset:
      name: offset
      in: query
      description: Index of the first element to return
      schema:
        type: integer
        default: 0
...
  schemas:
...
    intList:
      type: array
      items:
        type: integer
...
  responses:
    ServerError:
      description: An error occured while processing the request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: String representation of the exception
    InvalidRequest:
      description: Validation of input parameters failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                description: List of errors encountered during validation
                items:
                  type: string
    DatabaseError:
      description: The database query failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Description of the error
...
    NotFound:
      description: The requested resource could not be found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
...